changed newMonster to bool

+++ asciiquarium-sharp/Entity.cs
@@ -12,7 +12,7 @@ public class Entity
     public int Depth { get; set; }
     public int Physical { get; set; }
     public string? Color { get; set; }
-    public int[]? CallbackArgs { get; set; }
+    public double[]? CallbackArgs { get; set; }
     public TimeSpan? DieTime { get; set; }
     public string? DeathCb {get;set;}
 }
